+++ src/containers/exam/before-exam/login.tsx
@@ -17,7 +17,7 @@ import "react-simple-keyboard/build/css/index.css";
 import { useForceUpdate } from "../../../hooks";
 import { HeaderData } from "../future";
 import { FaUser, FaLock } from "react-icons/fa";
-import { InputFieldProps, HeaderProps, StepProps } from "../types";
+import { InputFieldProps, HeaderProps, StepProps } from "./types";
 
 function InputField({
   value,
@@ -100,7 +100,6 @@ export default function Login({ onNextStep }: StepProps) {
 
   const onSubmit = (e: any) => {
     e.preventDefault();
-    //TODO: more accurate validation?
     let okay = true;
     for (let input of inputs) {
       if (!input || input === "") {