+++ src/containers/exam/index.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { useTypedSelector } from "../../store/selector";
-import BeforeExam from "./before_exam";
-import MainExam from "./main_exam";
+import BeforeExam from "./before-exam";
+import MainExam from "./main-exam";
 
 export default function Exam() {
   const done = useTypedSelector((state) => state.beforeExamState);