+++ src/containers/exam/main-exam/types.ts
@@ -1,0 +1,29 @@
+export type TCandidateData = {
+  name: string;
+  avatar: string;
+  rollNumber: string;
+};
+
+export type TQuestionType = "mcq" | "numeric";
+
+// TODO: Reduce duplication in the data-model
+// @body Right now, most of the data is repeated unnecessarily. Maybe a better model exists. Replace this one with a model that's space efficient.
+export type TQuestion = {
+  id: string;
+  text: string;
+  type: TQuestionType;
+  choices?: string[];
+  correctMarks: number;
+  incorrectMarks: number;
+};
+
+export type TExamData = {
+  streamName: string;
+  examName: string;
+  subjects: string[];
+  sections: string[];
+  candidateData: TCandidateData;
+  questions: TQuestion[][];
+  totalTimeInMinutes: number;
+  calculatorAllowed: boolean;
+};