+++ src/store/types.ts
@@ -6,7 +6,25 @@ export type BeforeExamState = {
   step: number;
   done: boolean;
 };
-export type ExamState = {};
+
+export enum AnswerStatus {
+  MARKED_FOR_REVIEW,
+  MARKED_FOR_REVIEW_AND_ANSWERED,
+  ANSWERED,
+  NOT_ANSWERED,
+  NOT_VISITED,
+}
+export type AnswerState = {
+  index: number;
+  status: AnswerStatus;
+};
+
+export type ExamState = {
+  activeQuestionIndex: number;
+  activeSectionIndex: number;
+  activeSubjectIndex: number;
+  answers: AnswerState[];
+};
 
 export type GenericAction = {
   type: string;